==> notebooks/02_Inspect_Dataset.ipynb <==
# Inspect Dataset

Visualize downloaded images

import cv2
from ipywidgets import interactive, IntSlider
from IPython.display import clear_output
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
%matplotlib inline

# read image metadata
img_dir = '../data/raw'
metadata = pd.read_csv('../data/raw_image_metadata.csv', index_col=0)

metadata.head()

# view images one by one
def f(species_, class_, gender_, n_):
    # find the index
    ix = np.c_[
        metadata['species'] == species_,
        metadata['class'] == class_,
        metadata['gender'] == gender_,
        metadata['counter'] == n_
    ].all(1)
    fname = metadata[ix].fname.values[0]

    # read the image
    clear_output(wait=True)
    fig, ax = plt.subplots(figsize=(8, 8))
    Z = cv2.imread(fname)[:, :, ::-1]
    ax.imshow(Z)
    ax.set_title(fname)
    plt.show()
    

w = interactive(
    f,
    species_=metadata['species'].unique().tolist(),
    class_=metadata['class'].unique().tolist(),
    gender_=metadata['gender'].unique().tolist(),
    n_=IntSlider(min=metadata['counter'].min(), max=metadata['counter'].max(), step=1, value=0)
)
w

# view sheet of images
def f(species_, class_, gender_):
    # find the index
    ix = np.c_[
        metadata['species'] == species_,
        metadata['class'] == class_,
        metadata['gender'] == gender_
    ].all(1)
    fnames = metadata[ix].fname.values

    clear_output(wait=True)
    fig, axes = plt.subplots(10, 10, figsize=(20, 20))
    
    # read the images
    for i, fname in enumerate(fnames):
        ax = axes[i // 10, i % 10]
        # ax.set_title(fname)
        ax.set_xticks([])
        ax.set_yticks([])
        try:
            Z = cv2.imread(fname)[:, :, ::-1]
        except:
            continue

        ax.imshow(Z)

    plt.show()
    

w = interactive(
    f,
    species_=metadata['species'].unique().tolist(),
    class_=metadata['class'].unique().tolist(),
    gender_=metadata['gender'].unique().tolist()
)
w
